+++ CHANGELOG.md
@@ -4,6 +4,26 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [1.2.1] - 2025-06-14
+
+### Updated
+
+- Made duplicates useful within CI implementations
+
+---
+
+## [1.2.0] - 2025-06-14
+
+### Added
+
+- Detect duplicate or similar msgstr
+
+### Updated
+
+- Made translated text accessible via quick link
+
+---
+
 ## [1.1.0] - 2025-06-10
 
 ### Added